Description
In Python, arguments are values that are passed to functions, methods, or classes when they are called. Arguments can be used to provide additional information to the function, method, or class, or to change the way that it behaves. There are two types of arguments in Python: positional arguments and keyword arguments.