Description
We should know how to import a file or library in python  . Because while working  on data analysis we need to import libraries like numpy , pandas , matplotlib , csv file ( real time dataset ) etc.
So it would be beneficial  for us to know how to import a library. 
In this file simple math library is imported.