Description
A function is nothing more than a block of code. One of the best reasons to write Python functions is to separate reusable code from the core app so you don't repeat yourself. A classic example of a chunk of reusable code turned into a function is one that calculates the square of a number. 
This example  will help us to understand it better.